% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launch_rehab_app}
\alias{launch_rehab_app}
\title{Launch the PhysioRehab GUI in a browser}
\usage{
launch_rehab_app(...)
}
\arguments{
\item{...}{passed to [shiny::runApp].}
}
\description{
Launch the PhysioRehab GUI in a browser
}
